+++ crates/atlaspack/src/atlaspack.rs
@@ -9,8 +9,8 @@ use atlaspack_core::plugin::PluginContext;
 use atlaspack_core::plugin::PluginLogger;
 use atlaspack_core::plugin::PluginOptions;
 use atlaspack_core::types::AtlaspackOptions;
-use atlaspack_filesystem::os::OsFileSystem;
 use atlaspack_filesystem::FileSystemRef;
+use atlaspack_filesystem::OsFileSystem;
 use atlaspack_package_manager::NodePackageManager;
 use atlaspack_package_manager::PackageManagerRef;
 use atlaspack_plugin_rpc::RpcHostRef;
