+++ crates/atlaspack_filesystem/src/lib.rs
@@ -1,7 +1,8 @@
-pub mod file_system;
-pub mod memory;
-pub mod os;
+mod file_system;
+mod memory;
+mod os;
 pub mod utils;
 
 pub use self::file_system::*;
-pub use self::utils::*;
+pub use self::memory::*;
+pub use self::os::*;
