+++ crates/atlaspack_plugin_transformer_js/src/transformer.rs
@@ -214,7 +214,7 @@ mod test {
     plugin::PluginLogger,
     types::{Code, Dependency, Environment, Location, SourceLocation, SpecifierType, Symbol},
   };
-  use atlaspack_filesystem::memory::InMemoryFileSystem;
+  use atlaspack_filesystem::InMemoryFileSystem;
 
   use super::*;
