+++ packages/utils/node-resolver-rs/src/lib.rs
@@ -8,9 +8,9 @@ use bitflags::bitflags;
 use once_cell::unsync::OnceCell;
 
 pub use atlaspack_core::types::IncludeNodeModules;
-#[cfg(not(target_arch = "wasm32"))]
-pub use atlaspack_filesystem::os::OsFileSystem;
 pub use atlaspack_filesystem::FileSystem;
+#[cfg(not(target_arch = "wasm32"))]
+pub use atlaspack_filesystem::OsFileSystem;
 pub use cache::Cache;
 pub use cache::CacheCow;
 pub use error::ResolverError;
